e2e/pyproject.toml
[project]
name = "ptd-e2e"
version = "0.1.0"
description = "End-to-end tests for Posit Team Dedicated"
requires-python = ">=3.12"
dependencies = [
  "pytest>=8.3.5",
  "pytest-playwright>=0.6.2",
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --screenshot only-on-failure --full-page-screenshot --output test-results --junit-xml=test-results/junit.xml"
markers = [
  "e2e: end-to-end tests that require running services",
]
# Default base URL for e2e tests. Can be overridden via:
# - CLI: pytest --base-url https://example.com
# - ENV: PLAYWRIGHT_TEST_BASE_URL=https://example.com pytest
base_url = "http://localhost:8080"
